requirements.txt
# Hummus Tracker — Strait of Hormuz Oil Tanker Intelligence
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
websockets>=12.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
yfinance>=0.2.37
